Make datetime timezone more optional, not used for date
The datetime() sugar function previously could only accept the
timezone argument, because there was no good way to detect whether
the string was a timezone or not.  I decided it was good enough to
check for uppercase letters, because no DBIC column attribute will
have uppercase, and all timezones (except floating and local) have
at least one uppercase letter.

Also, there is no reason to check 'date' for a timezone argument.

Also lots of updates to documentation, and a test case for the
synopsis example.