Adjust implementation of json to not be generators
Previous implementation was to export methods specific to
the options used in the exporter use line, however I didn't
like the idea of needing generators for such a simple thing,
so changed it to just store the defaults in a per-package
variable.

This should also provide a bit more flexibility since the
defaults won't be accessed until the first call to the
json/jsonb methods, so defaults can change further even
after the symbol has been exported.