Improve the SchemaLoaderMixin to support inflate_json
This gives SchemaLoaderMixin the full understanding of whether the
generated package has inflate_json enabled package-wide or if it
needs enabled on individual columns.

I also added several features to make this possible, like generating
the verification package one and sharing it, selectively disabling
the -autoclean use mechanic, and letting the user override the use
line.